// clients/typescript/src/generated/types/schemaDataType.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from '@solana/kit';

export enum SchemaDataType {
  U8,
  U16,
  U32,
  U64,
  U128,
  I8,
  I16,
  I32,
  I64,
  I128,
  Bool,
  Char,
  String,
  VecU8,
  VecU16,
  VecU32,
  VecU64,
  VecU128,
  VecI8,
  VecI16,
  VecI32,
  VecI64,
  VecI128,
  VecBool,
  VecChar,
  VecString,
}

export type SchemaDataTypeArgs = SchemaDataType;

export function getSchemaDataTypeEncoder(): FixedSizeEncoder<SchemaDataTypeArgs> {
  return getEnumEncoder(SchemaDataType);
}

export function getSchemaDataTypeDecoder(): FixedSizeDecoder<SchemaDataType> {
  return getEnumDecoder(SchemaDataType);
}

export function getSchemaDataTypeCodec(): FixedSizeCodec<
  SchemaDataTypeArgs,
  SchemaDataType
> {
  return combineCodec(getSchemaDataTypeEncoder(), getSchemaDataTypeDecoder());
}